--- apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/enums/ApolloClientAuthenticationType.java
@@ -1,23 +1,0 @@
-package com.ctrip.framework.apollo.config.data.enums;
-
-/**
- * @author vdisk <vdisk@foxmail.com>
- */
-public enum ApolloClientAuthenticationType {
-
-  /**
-   * disable authentication
-   */
-  NONE,
-
-  /**
-   * enable oauth2 authentication
-   */
-  OAUTH2,
-
-  /**
-   * enable http-basic authentication
-   */
-  HTTP_BASIC,
-  ;
-}
