--- apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/enums/ApolloClientMessagingType.java
@@ -1,18 +1,0 @@
-package com.ctrip.framework.apollo.config.data.enums;
-
-/**
- * @author vdisk <vdisk@foxmail.com>
- */
-public enum ApolloClientMessagingType {
-
-  /**
-   * long polling
-   */
-  LONG_POLLING,
-
-  /**
-   * websocket
-   */
-  WEBSOCKET,
-  ;
-}
