+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/extension/enums/ApolloClientMessagingType.java
@@ -1,0 +1,18 @@
+package com.ctrip.framework.apollo.config.data.extension.enums;
+
+/**
+ * @author vdisk <vdisk@foxmail.com>
+ */
+public enum ApolloClientMessagingType {
+
+  /**
+   * long polling
+   */
+  LONG_POLLING,
+
+  /**
+   * websocket
+   */
+  WEBSOCKET,
+  ;
+}
