+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/extension/webclient/customizer/ApolloClientHttpBasicAuthenticationWebClientCustomizer.java
@@ -1,0 +1,23 @@
+package com.ctrip.framework.apollo.config.data.extension.webclient.customizer;
+
+import com.ctrip.framework.apollo.config.data.extension.webclient.filter.ApolloClientHttpBasicAuthenticationExchangeFilterFunction;
+import org.springframework.boot.web.reactive.function.client.WebClientCustomizer;
+import org.springframework.web.reactive.function.client.WebClient;
+
+/**
+ * @author vdisk <vdisk@foxmail.com>
+ */
+public class ApolloClientHttpBasicAuthenticationWebClientCustomizer implements WebClientCustomizer {
+
+  private final ApolloClientHttpBasicAuthenticationExchangeFilterFunction filterFunction;
+
+  public ApolloClientHttpBasicAuthenticationWebClientCustomizer(
+      ApolloClientHttpBasicAuthenticationExchangeFilterFunction filterFunction) {
+    this.filterFunction = filterFunction;
+  }
+
+  @Override
+  public void customize(WebClient.Builder webClientBuilder) {
+    webClientBuilder.filter(this.filterFunction);
+  }
+}
