+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/extension/webclient/customizer/ApolloClientOauth2ReactiveAuthenticationWebClientCustomizer.java
@@ -1,0 +1,24 @@
+package com.ctrip.framework.apollo.config.data.extension.webclient.customizer;
+
+import org.springframework.boot.web.reactive.function.client.WebClientCustomizer;
+import org.springframework.security.oauth2.client.web.reactive.function.client.ServerOAuth2AuthorizedClientExchangeFilterFunction;
+import org.springframework.web.reactive.function.client.WebClient;
+
+/**
+ * @author vdisk <vdisk@foxmail.com>
+ */
+public class ApolloClientOauth2ReactiveAuthenticationWebClientCustomizer implements
+    WebClientCustomizer {
+
+  private final ServerOAuth2AuthorizedClientExchangeFilterFunction filterFunction;
+
+  public ApolloClientOauth2ReactiveAuthenticationWebClientCustomizer(
+      ServerOAuth2AuthorizedClientExchangeFilterFunction filterFunction) {
+    this.filterFunction = filterFunction;
+  }
+
+  @Override
+  public void customize(WebClient.Builder webClientBuilder) {
+    webClientBuilder.filter(this.filterFunction);
+  }
+}
